+++ test-website/src/Pages/Test/Inputs/DateInputSection.tsx
@@ -17,8 +17,10 @@ interface DateInputSectionProps {
 }
 
 interface DateInputSectionState {
-    dateInput2Value: DateInputValue
     fieldValidity: FieldValidity
+    dateInput2Value: DateInputValue
+    dateInput7Value: DateInputValue
+    dateInput8Value: DateInputValue
 }
 
 export class DateInputSection extends React.Component<
@@ -27,7 +29,15 @@ export class DateInputSection extends React.Component<
 > {
     state: DateInputSectionState = {
         fieldValidity: {},
-        dateInput2Value: defaultDateInputValue
+        dateInput2Value: defaultDateInputValue,
+        dateInput7Value: dateInputValueFromMoment(moment(), {
+            includesTime: true,
+            timeZone: 'America/Los_Angeles'
+        }),
+        dateInput8Value: dateInputValueFromMoment(moment(), {
+            includesTime: true,
+            timeZone: 'America/Los_Angeles'
+        })
     }
 
     childValidChange = (fieldName: string, valid: boolean) => {
@@ -36,7 +46,12 @@ export class DateInputSection extends React.Component<
 
     render() {
         const { showValidation } = this.props
-        const { fieldValidity, dateInput2Value } = this.state
+        const {
+            fieldValidity,
+            dateInput2Value,
+            dateInput7Value,
+            dateInput8Value
+        } = this.state
 
         const vProps = { showValidation, onValidChange: this.childValidChange }
 
@@ -54,6 +69,7 @@ export class DateInputSection extends React.Component<
                         <DateInput
                             id={id}
                             name="dateInput0"
+                            timeZone="local"
                             validators={[]}
                             {...vProps}
                         />
@@ -64,6 +80,7 @@ export class DateInputSection extends React.Component<
                     <ValidityLabel valid={fieldValidity.dateInput1} />
                     <DateInput
                         name="dateInput1"
+                        timeZone="local"
                         validators={[DateValidators.required()]}
                         {...vProps}
                     />
@@ -75,6 +92,7 @@ export class DateInputSection extends React.Component<
                         <div className="mr-2">
                             <DateInput
                                 name="dateInput2"
+                                timeZone="local"
                                 value={dateInput2Value}
                                 onChange={dateInput2Value =>
                                     this.setState({ dateInput2Value })
@@ -98,7 +116,10 @@ export class DateInputSection extends React.Component<
                             onClick={() => {
                                 const m = moment('2001-01-01T10:00:00.000Z')
                                 this.setState({
-                                    dateInput2Value: dateInputValueFromMoment(m, false)
+                                    dateInput2Value: dateInputValueFromMoment(m, {
+                                        includesTime: false,
+                                        timeZone: 'local'
+                                    })
                                 })
                             }}
                         >
@@ -111,8 +132,9 @@ export class DateInputSection extends React.Component<
                     <ValidityLabel valid={fieldValidity.dateInput3} />
                     <DateInput
                         name="dateInput3"
+                        timeZone="local"
                         validators={[DateValidators.required({ includesTime: true })]}
-                        showTimeSelect
+                        includesTime
                         timeIntervals={10}
                         {...vProps}
                     />
@@ -122,6 +144,7 @@ export class DateInputSection extends React.Component<
                     <ValidityLabel valid={fieldValidity.dateInput4} />
                     <DateInput
                         name="dateInput4"
+                        timeZone="local"
                         validators={[DateValidators.required({ includesTime: true })]}
                         showPicker={false}
                         {...vProps}
@@ -132,11 +155,88 @@ export class DateInputSection extends React.Component<
                     <ValidityLabel valid={fieldValidity.dateInput5} />
                     <DateInput
                         name="dateInput5"
+                        timeZone="local"
                         readOnly
-                        defaultValue={dateInputValueFromMoment(moment(), false)}
+                        defaultValue={dateInputValueFromMoment(moment(), {
+                            includesTime: false,
+                            timeZone: 'local'
+                        })}
                         validators={[]}
                         {...vProps}
                     />
+                </div>
+                <div className="form-group">
+                    <label>
+                        Pacific time (uncontrolled), defaults to the current time
+                    </label>{' '}
+                    <ValidityLabel valid={fieldValidity.dateInput6} />
+                    <div className="d-flex align-items-baseline">
+                        <DateInput
+                            name="dateInput6"
+                            timeZone="America/Los_Angeles"
+                            defaultValue={dateInputValueFromMoment(moment(), {
+                                includesTime: true,
+                                timeZone: 'America/Los_Angeles'
+                            })}
+                            includesTime
+                            validators={[]}
+                            {...vProps}
+                        />
+                        <div className="ml-3 mr-5">Pacific</div>
+                    </div>
+                </div>
+                <div className="form-group">
+                    <label>Pacific time (controlled), defaults to the current time</label>{' '}
+                    <ValidityLabel valid={fieldValidity.dateInput7} />
+                    <div className="d-flex align-items-baseline">
+                        <DateInput
+                            name="dateInput7"
+                            timeZone="America/Los_Angeles"
+                            value={dateInput7Value}
+                            onChange={dateInput7Value =>
+                                this.setState({ dateInput7Value })
+                            }
+                            includesTime
+                            validators={[]}
+                            {...vProps}
+                        />
+                        <div className="ml-3 mr-5">Pacific</div>
+                        <div>
+                            UTC:{' '}
+                            <b>
+                                {dateInput7Value.moment &&
+                                    dateInput7Value.moment.utc().format('M/D/YYYY H:mm')}
+                            </b>
+                        </div>
+                    </div>
+                </div>
+                <div className="form-group">
+                    <label>
+                        Pacific time (controlled), no picker, defaults to the current time
+                    </label>{' '}
+                    <ValidityLabel valid={fieldValidity.dateInput8} />
+                    <div className="d-flex align-items-baseline">
+                        <DateInput
+                            name="dateInput8"
+                            timeZone="America/Los_Angeles"
+                            value={dateInput8Value}
+                            onChange={dateInput8Value =>
+                                this.setState({ dateInput8Value })
+                            }
+                            showPicker={false}
+                            includesTime
+                            validators={[]}
+                            {...vProps}
+                        />
+                        <div className="ml-3 mr-5">Pacific</div>
+                        <div>
+                            UTC:{' '}
+                            <b>
+                                {dateInput8Value.moment &&
+                                    dateInput8Value.moment.utc().format('M/D/YYYY H:mm')}
+                            </b>
+                        </div>
+                    </div>
                 </div>
             </div>
         )